Update CommitBot.py
Reference https://github.com/commitsto/commits.to/issues/240

Updated MENTION_REGEX so it will match allowed URL characters (and only those in allowable); so people don’t get into bad habits early.

Added two different options for making commits with the bot. The statement /iwill or the statement /commitsto. Because, well, why not?

It is way too early and I’m not used to the time change and the word commit is reminding me of the old movie Rocketman.